Use translators in MessagePair (#44)
* Add translators for std msgs

* Fix includes

* Update MessagePair with clock

* Add exceptions

+++ source/modulo_new_core/include/modulo_new_core/communication/MessageType.hpp
@@ -3,7 +3,7 @@
 namespace modulo_new_core::communication {
 
 enum class MessageType {
-  BOOL, FLOAT64, FLOAT64_MULTI_ARRAY, INT64, STRING
+  BOOL, FLOAT64, FLOAT64_MULTI_ARRAY, INT32, STRING
 };
 
 }// namespace modulo_new_core::communication
